homework4

+++ src/homework4/operations/Division.java
@@ -2,7 +2,7 @@ package homework4.operations;
 
 public class Division implements Operation {
     @Override
-    public float execute(float a, float b) {
+    public double execute(double a, double b) {
         return a / b;
     }
 }
